pyproject.toml
[tool.poetry]
name = "cpudiscordbot"
version = "0.1.0"
description = ""
authors = ["Ethan Chapman <ethan.chapman0@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
"discord.py" = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
